Extract daily location and subscriber location subset tests & convert to pytest
