+++ flowmachine/tests/test_query_subscribers.py
@@ -478,157 +478,6 @@ class test_aggregate_locations(TestCase):
             self.assertFalse(w)
 
 
-class test_dailylocations(TestCase):
-    """
-    Test for daily locations
-    """
-
-    def setUp(self):
-        """
-        Method that sets-up the test environment.
-        Here we are also creating tables at every
-        test and deleting it immediately after.
-        """
-
-        self.engine = Query.connection.engine
-
-        self._create_test_schema()
-
-    def _create_test_schema(self):
-        """
-        Private method that creates a test schema.
-        """
-        q = """
-            CREATE SCHEMA IF NOT EXISTS daily_locations;
-        """
-        self.engine.execute(q)
-        self.engine.commit()
-
-    def _drop_test_schema(self):
-        """
-        Private method for dropping schema.
-        """
-        q = """
-            DROP SCHEMA IF EXISTS daily_locations CASCADE;
-        """
-        self.engine.execute(q)
-        self.engine.commit()
-
-    def tearDown(self):
-        """
-        Method that drops test schema.
-        """
-        self._drop_test_schema()
-
-    def test_can_return_df(self):
-        """
-        daily_location() can be initialised and return a dataframe.
-        """
-
-        dl = daily_location("2016-01-01")
-        self.assertIs(type(dl.get_dataframe()), pd.DataFrame)
-
-    def test_equivalent_to_locate_subscribers(self):
-        """
-        daily_location() is equivalent to the MostFrequentLocation().
-        """
-        mfl = MostFrequentLocation("2016-01-01", "2016-01-02")
-        mfl_df = mfl.get_dataframe()
-
-        dl = daily_location("2016-01-01", method="most-common")
-        dl_df = dl.get_dataframe()
-
-        self.assertTrue((dl_df == mfl_df).all().all())
-
-    def test_equivalent_to_locate_subscribers_with_time(self):
-        """
-        daily_location() is equivalent to the MostFrequentLocation() with timestamps.
-        """
-        mfl = MostFrequentLocation("2016-01-01 18:00:00", "2016-01-02 06:00:00")
-        mfl_df = mfl.get_dataframe()
-
-        dl = daily_location(
-            "2016-01-01 18:00:00", stop="2016-01-02 06:00:00", method="most-common"
-        )
-        dl_df = dl.get_dataframe()
-
-        self.assertTrue((dl_df == mfl_df).all().all())
-
-    def test_can_be_stored(self):
-        """
-        We can store a daily_location() in the right place.
-        """
-
-        schema = "cache"
-        dl = daily_location("2016-01-01", level="cell")
-        table_name = dl.table_name.split(".")[1]
-        dl.store().result()
-        self.assertTrue(dl.connection.has_table(table_name, schema=schema))
-        dl = daily_location("2016-01-01", level="cell")
-        self.assertTrue(table_name in dl.get_query())
-        dl.invalidate_db_cache()
-
-    def test_storing_updates(self):
-        """
-        Storing updates the query after storing.
-        """
-
-        dl = daily_location("2016-01-01", level="cell")
-        dl.store().result()
-        sql = dl.get_query()
-        self.assertTrue(sql.startswith("SELECT * FROM cache"))
-        dl.invalidate_db_cache()
-
-    def test_force(self):
-        """
-        We can force an overwrite of the table.
-        """
-        dl = daily_location("2016-01-01", level="cell")
-        dl.store().result()
-        sql = "DELETE FROM {}".format(dl.table_name)
-        dl.connection.engine.execute(sql)
-        # This should not raise an error
-        dl.store().result()
-        self.assertTrue(len(dl) == 0)
-        dl.store(force=True)
-        self.assertTrue(len(dl.head()) != 0)
-        dl.invalidate_db_cache()
-
-    def test_non_standard_params(self):
-        """
-        Non-standard parameters don't trigger reading the cache.
-        """
-        dl = daily_location("2016-01-01", level="cell")
-        table_name = dl.table_name
-        dl.store().result()
-        dl2 = daily_location("2016-01-01", level="cell", method="most-common")
-        self.assertFalse(table_name in dl2.get_query())
-        dl.invalidate_db_cache()
-
-    def test_works_with_pcods(self):
-        """
-        We can get daily locations with p-codes rather than the standard names.
-        """
-
-        dl = daily_location("2016-01-05", level="admin3", column_name="admin3pcod")
-        df = dl.get_dataframe()
-        self.assertTrue(df.admin3pcod[0].startswith("524"))
-
-    def test_hours(self):
-        """
-        Test that daily locations handles the hours parameter
-        """
-
-        # Lower level test test that subsetdates handles this correctly
-        # we're just testing that it is passed on in this case.
-
-        dl1 = daily_location("2016-01-01", level="cell")
-        dl2 = daily_location("2016-01-01", level="cell", hours=(19, 23))
-        dl3 = daily_location("2016-01-01", level="cell", hours=(19, 20))
-
-        self.assertTrue(len(dl1) > len(dl2) > len(dl3))
-
-
 class test_spanned_home_locations(TestCase):
     def setUp(self):
         self.dls = [
@@ -681,72 +530,3 @@ class test_home_locations(TestCase):
         self.assertEquals(self.hdf.ix["E1n7JoqxPBjvR5Ve"][0], "Baglung")
         self.assertEquals(self.hdf.ix["gkBLe0mN5j3qmRpX"][0], "Myagdi")
         self.assertEquals(self.hdf.ix["5Kgwy8Gp6DlN3Eq9"][0], "Kapilbastu")
-
-
-class test_missing_dates(TestCase):
-    """
-    Tests that deal with the logic for dealing with missing dates.
-    """
-
-    def test_daily_locs_errors(self):
-        """
-        daily_location() errors when we ask for a date that does not exist.
-        """
-
-        with self.assertRaises(MissingDateError):
-            dl = daily_location("2016-01-31")
-
-
-def test_subscriber_location_subset_column_names(exemplar_level_param):
-    ss = SubscriberLocationSubset(
-        "2016-01-01", "2016-01-07", min_calls=1, **exemplar_level_param
-    )
-    assert ss.head(0).columns.tolist() == ss.column_names
-
-
-class test_subscriber_location_subset(TestCase):
-    def test_subscribers_make_atleast_one_call_in_admin0(self):
-        """
-        The set of subscribers who make at least one call within admin0 over
-        whole test time period should be equal to set of unique subscribers
-        in test calls table.
-        """
-
-        start, stop = "2016-01-01", "2016-01-07"
-
-        sls = SubscriberLocationSubset(start, stop, min_calls=1, level="admin0")
-        us = UniqueSubscribers(start, stop, table="events.calls")
-
-        sls_subs = set(sls.get_dataframe()["subscriber"])
-        us_subs = set(us.get_dataframe()["subscriber"])
-
-        self.assertEquals(sls_subs, us_subs)
-
-    def test_subscribers_who_make_atleast_3_calls_in_central_development_region(self):
-        """
-        Test that we can find subsets for multiple geometries at same time. Will
-        find subscribers who have made at least 2 calls in any of the admin2 regions
-        within Central Development admin1 region.
-        """
-        start, stop = "2016-01-01", "2016-01-07"
-        regions = Table("admin2", "geography").subset(
-            "admin1name", ["Central Development Region"]
-        )
-
-        sls = SubscriberLocationSubset(
-            start,
-            stop,
-            min_calls=2,
-            level="polygon",
-            column_name="admin2pcod",
-            polygon_table=regions,
-        )
-
-        df = sls.get_dataframe()
-
-        # we have results for multiple regions
-        self.assertGreater(len(df.admin2pcod.unique()), 1)
-
-        # some users should have have made at least 2 calls in more than one region
-        # and should therefore appear twice
-        self.assertGreater(len(df[df.duplicated("subscriber")]), 0)
